requirements.txt
Flask>=2.0.1
Werkzeug>=2.2.2
python-dotenv>=0.19.0
pymongo>=3.11.4
pymongo[srv]